+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Added the `test_local_timezone()` and `set_local_timezone()` helpers to ease testing.
 - Added support of ISO 8601 duration parsing.
 - Added support of ISO 8601 interval parsing.
+- Added a `local()` helper.
 
 ### Changed